requirements.txt
git+https://github.com/python-telegram-bot/ptbcontrib.git
Telethon==1.23.0
python_telegram_bot==13.7
SQLAlchemy==1.4.26
psycopg2-binary